procwatch: Move CommandStream to new module
Move CommandStream from the utils junkyard to the new procwatch module
and the tests to a new test module, with minimal changes to keep it
working.
