+++ lib/vdsm/utils.py
@@ -35,7 +35,6 @@ import distutils.spawn
 import errno
 import functools
 import glob
-import io
 import logging
 import six
 import sys
@@ -278,60 +277,6 @@ def NoIntrPoll(pollfun, timeout=-1):
 
         if endtime is not None:
             timeout = max(0, endtime - monotonic_time())
-
-
-class CommandStream(object):
-    def __init__(self, command, stdoutcb, stderrcb):
-        self._command = command
-        self._poll = select.epoll()
-        self._iocb = {}
-
-        # In case both stderr and stdout are using the same fd the
-        # output is squashed to the stdout (given the order of the
-        # entries in the dictionary)
-        self._iocb[self._command.stderr.fileno()] = stderrcb
-        self._iocb[self._command.stdout.fileno()] = stdoutcb
-
-        for fd in self._iocb:
-            self._poll.register(fd, select.EPOLLIN)
-
-    def _poll_input(self, fileno):
-        self._iocb[fileno](os.read(fileno, io.DEFAULT_BUFFER_SIZE))
-
-    def _poll_event(self, fileno):
-        self._poll.unregister(fileno)
-        del self._iocb[fileno]
-
-    def _poll_timeout(self, timeout):
-        fdevents = NoIntrPoll(self._poll.poll, timeout)
-
-        for fileno, event in fdevents:
-            if event & select.EPOLLIN:
-                self._poll_input(fileno)
-            elif event & (select.EPOLLHUP | select.EPOLLERR):
-                self._poll_event(fileno)
-
-    @property
-    def closed(self):
-        return len(self._iocb) == 0
-
-    def receive(self, timeout=None):
-        """
-        Receiving data from the command can raise OSError
-        exceptions as described in read(2).
-        """
-        if timeout is None:
-            poll_remaining = -1
-        else:
-            endtime = monotonic_time() + timeout
-
-        while not self.closed:
-            if timeout is not None:
-                poll_remaining = endtime - monotonic_time()
-                if poll_remaining <= 0:
-                    break
-
-            self._poll_timeout(poll_remaining)
 
 
 def stripNewLines(lines):
