+++ tests/utilsTests.py
@@ -22,12 +22,10 @@ from __future__ import print_function
 import collections
 import contextlib
 import copy
-import cpopen
 import errno
 import fcntl
 import gc
 import logging
-import operator
 import os
 import os.path
 import select
@@ -772,109 +770,6 @@ class RoundTests(TestCaseBase):
     ])
     def test_round(self, n, size, result):
         self.assertEqual(utils.round(n, size), result)
-
-
-@expandPermutations
-class CommandStreamTests(TestCaseBase):
-
-    def assertUnexpectedCall(self, data):
-        raise AssertionError("Unexpected data: %r" % data)
-
-    def _startCommand(self, command):
-        return cpopen.CPopen(command)
-
-    @permutations([
-        (['echo', '-n', '%s'], True, False),
-        (['sh', '-c', 'echo -n "%s" >&2'], False, True),
-    ])
-    def test_receive(self, cmd, recv_out, recv_err):
-        text = bytes('Hello World')
-        received = bytearray()
-
-        def recv_data(buffer):
-            # cannot use received += buffer with a variable
-            # defined in the parent function.
-            operator.iadd(received, buffer)
-
-        cmd[-1] = cmd[-1] % text
-
-        c = self._startCommand(cmd)
-        p = utils.CommandStream(
-            c,
-            recv_data if recv_out else self.assertUnexpectedCall,
-            recv_data if recv_err else self.assertUnexpectedCall)
-
-        while not p.closed:
-            p.receive()
-
-        retcode = c.wait()
-
-        self.assertEqual(retcode, 0)
-        self.assertEqual(text, received)
-
-    @permutations([
-        (['cat'], True, False),
-        (['sh', '-c', 'cat >&2'], False, True),
-    ])
-    def test_write(self, cmd, recv_out, recv_err):
-        text = bytes('Hello World')
-        received = bytearray()
-
-        def recv_data(buffer):
-            # cannot use received += buffer with a variable
-            # defined in the parent function.
-            operator.iadd(received, buffer)
-
-        c = self._startCommand(cmd)
-        p = utils.CommandStream(
-            c,
-            recv_data if recv_out else self.assertUnexpectedCall,
-            recv_data if recv_err else self.assertUnexpectedCall)
-
-        c.stdin.write(text)
-        c.stdin.flush()
-        c.stdin.close()
-
-        while not p.closed:
-            p.receive()
-
-        retcode = c.wait()
-
-        self.assertEqual(retcode, 0)
-        self.assertEqual(text, str(received))
-
-    def test_timeout(self):
-        c = self._startCommand(["sleep", "5"])
-        p = utils.CommandStream(c, self.assertUnexpectedCall,
-                                self.assertUnexpectedCall)
-
-        with self.assertElapsed(2):
-            p.receive(2)
-
-        self.assertEqual(p.closed, False)
-
-        c.terminate()
-
-        self.assertEqual(c.wait(), -signal.SIGTERM)
-
-    @permutations((
-        ('kill', -signal.SIGKILL),
-        ('terminate', -signal.SIGTERM),
-    ))
-    def test_signals(self, method, expected_retcode):
-        c = self._startCommand(["sleep", "2"])
-        p = utils.CommandStream(c, self.assertUnexpectedCall,
-                                self.assertUnexpectedCall)
-
-        getattr(c, method)()
-
-        try:
-            with self.assertElapsed(0):
-                p.receive(2)
-        finally:
-            retcode = c.wait()
-
-        self.assertEqual(retcode, expected_retcode)
 
 
 class FakeLogger(object):
